requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Jinja2==3.1.6
PyMySQL==1.1.1
SQLAlchemy==2.0.40
Werkzeug==3.1.3
python-dotenv==1.1.0
WTForms==3.2.1
blinker==1.9.0